Move regression funcs into utils.utils.
