+++ xcp_d/utils/utils.py
@@ -7,7 +7,8 @@ import os
 import nibabel as nb
 import numpy as np
 from pkg_resources import resource_filename as pkgrf
-from scipy.signal import butter, filtfilt
+from scipy.signal import butter, detrend, filtfilt
+from sklearn.linear_model import LinearRegression
 from templateflow.api import get as get_template
 
 
@@ -427,3 +428,47 @@ def butter_bandpass(data, fs, lowpass, highpass, order=2):
                                         padlen=3 * (max(len(b), len(a)) - 1))
 
     return filtered_data
+
+
+def linear_regression(data, confound):
+    """Perform linear regression with sklearn's LinearRegression.
+
+    Parameters
+    ----------
+    data : numpy.ndarray
+        vertices by timepoints for bold file
+    confound : numpy.ndarray
+       nuisance regressors - vertices by timepoints for confounds matrix
+
+    Returns
+    -------
+    numpy.ndarray
+        residual matrix after regression
+    """
+    regression = LinearRegression(n_jobs=1)
+    regression.fit(confound.T, data.T)
+    y_predicted = regression.predict(confound.T)
+
+    return data - y_predicted.T
+
+
+def demean_detrend_data(data):
+    """Mean-center and remove linear trends over time from data.
+
+    Parameters
+    ----------
+    data : numpy.ndarray
+        vertices by timepoints for bold file
+
+    Returns
+    -------
+    detrended : numpy.ndarray
+        demeaned and detrended data
+    """
+    demeaned = detrend(data, axis=- 1, type='constant', bp=0,
+                       overwrite_data=False)  # Demean data using "constant" detrend,
+    # which subtracts mean
+    detrended = detrend(demeaned, axis=- 1, type='linear', bp=0,
+                        overwrite_data=False)  # Detrend data using linear method
+
+    return detrended  # Subtract these predicted values from the demeaned data
